Revert "feat: kanban of main page"
