+++ frontend/gfi-bot-web/src/style/gfiStyle.css
@@ -122,96 +122,3 @@
     width: -webkit-max-content;
     width: -moz-fit-content;
 }
-
-.gfi-wrapper {
-    display: flex;
-    border-radius: 10px;
-    box-sizing: border-box;
-    border: 2px white solid;
-    width: 100%;
-    height: 100%;
-    padding-right:  5px;
-    padding-left: 5px;
-    color: white;
-}
-
-.gfi-wrapper.kanban {
-    flex-flow: column wrap;
-    align-items: center;
-}
-
-.kanban.wrapper {
-    width: 90%;
-    display: flex;
-    flex-flow: row-reverse wrap;
-    justify-content: space-evenly;
-}
-
-.kanban.wrapper > .kanban div {
-    display: flex;
-    flex-flow: column wrap;
-    align-items: center;
-}
-
-.kanban.data > div:first-child {
-    font-size: 1.2em;
-    font-weight: bold;
-}
-
-.kanban.data > div:last-child {
-    font-size: 1.4em;
-}
-
-.kanban.border {
-    background-color: white;
-    width: 1px;
-    border-radius: 999px;
-    margin: 5px;
-}
-
-.gfi-wrapper.tags {
-    display: flex;
-    flex-flow: column wrap;
-    width: 95%;
-    margin-bottom: 10px;
-    padding: 5px 10px;
-}
-
-.tags.wrapper {
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: flex-start;
-    align-items: stretch;
-}
-
-.gfi-wrapper.tags div:first-child {
-    font-size: large;
-    font-weight: bold;
-}
-
-.tags > button {
-    margin: 3px 2px;
-    padding-left: 10px;
-    padding-right: 10px;
-}
-
-button.gfi-rounded  {
-    border: white 1px solid;
-    border-radius: 999px;
-    background-color: transparent;
-    box-sizing: border-box;
-    color: white;
-    transition: 0.1s ease-in-out;
-}
-
-button.gfi-rounded:hover {
-    color: black;
-    background-color: white;
-    transition: 0.1s ease-in-out;
-}
-
-button.gfi-rounded:active {
-    color: white;
-    background-color: cornflowerblue;
-    transition: 0.1s ease-in-out;
-}
